refactor: consolidate roadmap services
