+++ src/services/roadmapService.js
@@ -1,5 +1,6 @@
 // services/roadmapService.js
-import { API_BASE_URL } from '../config.js';
+import { useRef } from 'react';
+import { API_BASE_URL, WS_BASE_URL } from '../config.js';
 
 export const assignRoadmapToUser = async (roadmapId, userId, authToken) => {
   try {
@@ -126,3 +127,97 @@ export const fetchUserRoadmaps = async (userId, authToken) => {
     return null;
   }
 };
+
+export function useRoadmapWebSocket({ onMessage } = {}) {
+  const wsRef = useRef(null);
+
+  const connect = () => {
+    if (wsRef.current) return; // avoid reconnecting if already connected
+
+    const ws = new WebSocket(`${WS_BASE_URL}/ws/roadmap`);
+    wsRef.current = ws;
+
+    ws.onmessage = (event) => {
+      try {
+        const data = JSON.parse(event.data);
+
+        if (data?.chat_id) {
+          try {
+            localStorage.setItem('chatRoadmapId', data.chat_id);
+          } catch (e) {
+            console.warn('Failed to save chatRoadmapId to localStorage:', e);
+          }
+          return;
+        }
+
+        if (data.token !== undefined) {
+          onMessage?.(data.token);
+        } else {
+          onMessage?.(data);
+        }
+      } catch (error) {
+        console.warn('Failed to parse WebSocket message:', error);
+        onMessage?.(event.data);
+      }
+    };
+
+    ws.onopen = () => {
+      console.log('[Roadmap WebSocket Connected]');
+    };
+
+    ws.onclose = () => {
+      console.log('[Roadmap WebSocket Closed]');
+      wsRef.current = null;
+    };
+
+    ws.onerror = (error) => {
+      console.error('[Roadmap WebSocket Error]:', error);
+    };
+  };
+
+  const close = () => {
+    if (wsRef.current) {
+      wsRef.current.close();
+      wsRef.current = null;
+    }
+  };
+
+  const sendMessage = (data) => {
+    try {
+      if (wsRef.current?.readyState === WebSocket.OPEN) {
+        const chatRoadmapId = localStorage.getItem('chatRoadmapId');
+        const payload = chatRoadmapId
+          ? { ...data, chat_id: parseInt(chatRoadmapId, 10) }
+          : data;
+
+        wsRef.current.send(JSON.stringify(payload));
+      } else {
+        console.warn('WebSocket is not open. Ready state:', wsRef.current?.readyState);
+      }
+    } catch (error) {
+      console.error('Error sending WebSocket message:', error);
+    }
+  };
+
+  return { connect, sendMessage, close };
+}
+
+export async function fetchRoadmapMessages(chatRoadmapId) {
+  const token = localStorage.getItem('token');
+  if (!chatRoadmapId) return [];
+
+  const headers = token ? { Authorization: token } : {};
+
+  try {
+    const res = await fetch(`${API_BASE_URL}/messages/${chatRoadmapId}`, { headers });
+    if (!res.ok) throw new Error('Failed to fetch messages');
+    const data = await res.json();
+    return (data.messages || []).map((m) => ({
+      role: m.message_from === 'assistant' ? 'agent' : 'user',
+      text: m.text,
+    }));
+  } catch (err) {
+    console.error('Error fetching roadmap messages:', err);
+    return [];
+  }
+}
